Fraction->double in SdrObject::Resize
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/201733
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202759
Tested-by: Tomaž Vajngerl <quikee@gmail.com>

+++ svx/source/svdraw/svdfmtf.cxx
@@ -450,7 +450,7 @@ void ImpSdrGDIMetaFileImport::InsertObj(SdrObject* pObj1, bool bScale)
     {
         if(mbSize)
         {
-            pObj->NbcResize(Point(), maScaleX, maScaleY);
+            pObj->NbcResize(Point(), double(maScaleX), double(maScaleY));
         }
 
         if(mbMov)
