+++ svx/source/svdraw/svdpdf.cxx
@@ -682,7 +682,7 @@ void ImpSdrPdfImport::InsertObj(SdrObject* pObj1, bool bScale)
     {
         if (mbSize)
         {
-            pObj->NbcResize(Point(), maScaleX, maScaleY);
+            pObj->NbcResize(Point(), double(maScaleX), double(maScaleY));
         }
 
         if (mbMov)
